% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcd_majority.R
\name{nlcd_majority}
\alias{nlcd_majority}
\title{Calculate majority NLCD per polygon and update NAs}
\usage{
nlcd_majority(nlcd, DMA_dir, DMA_file)
}
\arguments{
\item{nlcd}{This is the file path to the nlcd raster}

\item{DMA_dir}{This is the file path to where the GIS file of version 1 DMA assignments for a county}

\item{DMA_file}{This is the file name of the GIS file of version 1 DMA assignment for a county}
}
\value{
DMA_1
}
\description{
Calculate majority NLCD per polygon and update NAs
}
\examples{
DMA_v1 <- nlcd_majority(nlcd="C:/Users/ecostel/DMA_Mapping/DMA_Mapping/GIS/NLCD_2016_Land_Cover_L48_20190424.img", DMA_dir = "C:/Users/ecostel/DMA_Mapping/DMA_Mapping/GIS", DMA_file = "Jefferson_DMAs_2019-2E" )
}
