% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_dma.R
\name{assign_dma}
\alias{assign_dma}
\title{Assign DMA to polygons}
\usage{
assign_dma(DMA, DMA_dir, countyname, gaps, tribal, zcode, pubyear)
}
\arguments{
\item{DMA}{This is the DMA R object output by the nlcd_majority() function. This can also be the file name of the county DMA GIS file}

\item{DMA_dir}{IF reading in a GIS file for param DMA, this is the file path of the county DMA GIS file}

\item{countyname}{name of county with DMAs being updated by this function}

\item{gaps}{TRUE/FAlSE indicating whether there are gaps in the county tax lot GIS file}

\item{tribal}{TRUE/FALSe indicating whether there are tribal areas within the county}

\item{zcode}{TRUE/FALSE indicating whether zoning data is included for this county in the statwide basedata layer}

\item{pubyear}{2016 or 2019 indicating the year of the public land management GIS base data layer}
}
\value{
DMA
}
\description{
Assign DMA to polygons
}
\examples{
DMA_v2 <- assign_dma(DMA= DMA, DMA_dir, countyname = "Jefferson County", gaps= FALSE, tribal = TRUE, zcode = TRUE, pubyear= 2019)
}
